Btrfs-progs: fix restore to fall back to the broken open_ctree
We don't need most of the roots when doing restore, like the extent tree.  So if
the recovery open_ctree fails because it's trying to open one of these useless
roots just fall back to open_ctree_broken.  This will just open the chunk root
which is the bare minimum of what we need to operate.  Then from there we can
setup the tree_root and the fs_root, and if either of those are gone we can't
restore anyway.  Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>