btrfs-progs: add a metadata usage internal command
We had a problem recently where some unknown application caused every
developers fs to explode in metadata usage.  I needed a tool to easily
work out where the metadata usage was, so I wrote this tool.  It works
similarly to du, it'll traverse down the given directory and give you
the metadata size of all the elements under the tree in that root.  It
won't cross into different roots, and it spits out a bunch of different
statistics about the metadata usage.  You can use the -s option to get a
more du-like output.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>