btrfs-progs: validate all trees before normal checking
Currently btrfsck falls apart if there's major corruption in any of the
trees that would stop btrfs_search_slot() from succeeding.  The sort of
failures are things like csum mismatches, pointing at tree blocks that
don't belong to the current tree, or any failure from
btrfs_check_node/btrfs_check_leaf.

Some of these failures we could probably recover from, but as a first
pass simply remove links to nodes/leafs that fail these basic checks.
For root items that point at invalid nodes/leafs we either reinit in the
case of a global tree, or remove the root item for fs trees.

Once we're sure we aren't going to fail to do normal searches we can
continue with the rest of fsck.  The repair will be able to fix anything
we remove from the extent tree and most things we lose in the fs tree.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>