btrfs-progs: fix infinite loop in find_normal_file_extent
This function is supposed to see if there's a regular extent for a file
to determine if it is a normal file or not.  However we never advance
the path at all, we just spin on the same slot.  Fix the loop up to
increment the slot and move to the next leaf if necessary.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>