btrfs-progs: always exclude referenced blocks when repairing
Currently we have a hodge-podge set of mechanisms to avoid allocating
over referenced blocks in the file system when repairing the file
system.  This is problematic for a few reasons

1. We can clear the inode cache without excluding anything, so we could
   overwrite metadata that we may reference elsewhere.  This isn't
   likely, but still is a gap.
2. Fsck may not correctly correct something, and then the user is doubly
   screwed because we could have overwritten the file system and thus
   have no way to get back to the original broken state.
3. We duplicate this code in a few places, one for init'ing a new extent
   tree, one for lowmem, and one for the normal case.

Address these issues by excluding all of the blocks that we reference
currently at the repair time.  Then rip out all of the duplicate places
that we do exclusion, as it's now covered by the main function.  The
drawback is we could have ENOSPC where we wouldn't have before because
we're pinning everything down, but we can address that in a general way
in the future if that does turn out to be a real problem.  The problem
would still exist in the current code to some degree, but less so in the
normal mode because we un-exclude blocks as we process them, potentially
opening up space.  In lowmem mode it's less of a problem because it will
force a chunk allocation if necessary.  However the normal allocator
should do that if necessary.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>