btrfs-progs: mkfs: make per-bg free space root
Now that we have the supporting code in place, allow mkfs to make the
per-bg free space root.  This mostly happens automagically, we just have
to set the root key offset to the block group bytenr, and make sure to
recow the roots so they get moved out of the temporary system chunk.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>