btrfs-progs: make btrfs_create_free_space_tree work with extent tree v2
Extent tree v2 no longer tracks all allocated blocks in the extent tree,
so we need to get all the used blocks by walking the metadata blocks, as
well as walking the per-block group extent trees.  Use the helper we
have for this, and then generate the free space tree based on the gaps
between the used areas in the used extent_io_tree.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>