btrfs-progs: check: handle init csum root with extent tree v2
If we have extent tree v2 enabled then we have per-block group roots
that we need to iterate through to fill the csums.  The other trick is
that we can have overlapping ranges of extent ranges with extent tree
v2, so keep track of what we've already csum'ed so we don't re-csum a
range.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>