btrfs-progs: make btrfs_prev_extent_item extent tree v2 aware
We can have overlapping extent items in the extent tree for extent tree
v2.  This needs to be handled with btrfs_prev_extent_item, as the start
could be < the bytenr we're searching, but could overlap the range, so
we need to account for this if we have extent tree v2 enabled.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>