btrfs-progs: handle freeing extents for extent tree v2
With extent tree v2 we will track references of actually referenced file
extents, which means that we could have overlapping ranges.  Handle this
by checking if we made any holes by removing our extent item, and if so
free'ing the appropriate amount.

For metadata we don't track the space in the extent root, so simply add
the space to the free space tree and do the appropriate accounting.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>