btrfs-progs: add free space for new block groups
With extent tree v2 we are going to be creating per-block group roots at
block group creation time, however because we won't yet have the free
space tree root at that point we won't be able to cache the block group.
Fix this by adding the free space cache for new block groups like we do
in the kernel, this way we can immediately allocate blocks from the
block group while we're creating the file system.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>