btrfs-progs: track the bg and remap root in the super
Because we can remap logical addresses to different block groups we have
to be able to read the block group roots and remap roots before we load
any other metadata blocks, as they could be relocated and we need to
know where to find them.

This patch adds the related information to the super block so we know
where to find these two roots, and then adds the code to write them to
the backup supers.  Since we're getting rid of the global csum and
extent roots in extent tree v2 we can simply use these slots for the
remap and block group roots in the backup roots array.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>