btrfs-progs: add new COW rules for extent tree v2
All of the previous COW rules will apply, but with extent tree v2 we add
a new COW rule where we must COW if the snapshot id of the buffer is
less than the current snapshot id of the root.  If the root's snapshot
id has been increased we know that the block is now shared by a new root
and we must cow this block.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>